refactor(core): update imports to use @/ instead of ./
