+++ core/monorule/src/rules/type.ts
@@ -1,0 +1,4 @@
+import type { defineRule } from "./define";
+
+export type Rule = ReturnType<typeof defineRule>;
+export type RuleMap = Record<string, Rule>;
